feat: Add flash-attn and liger-kernel dependencies (#11)
* feat: Upgrade to Python 3.12 and update dependencies

The project now requires Python 3.12 as the minimum version. The huggingface-cli dependency in README.md was also updated to hf.  A uv.lock file was created.

* feat: Add flash-attn and liger-kernel dependencies

This commit introduces `flash-attn` and `liger-kernel` as new dependencies to the project.
It also updates the `torchrun` command in `README.md` and the `uv.lock` file.

* fix: Handle missing qwen_vl_utils dependency in BaseDataset

* feat: Add qwen-vl-utils dependency and improve config loading

This commit introduces the `qwen-vl-utils` as a dependency and improves config loading by searching in project root.
It also updates example paths and number of processes.

* feat: Add qwen-vl-utils dependency and improve config loading